# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zerizeha API
  version: "1.0"
  description: Zerizeha backend API
servers:
  - url: /
paths:
...
  /api/auth/google/callback:
    get:
      operationId: GoogleCallback
      summary: google oauth callback
      description: exchanges code, returns tokens
      tags:
        - Auth
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
  /api/auth/github/callback:
    get:
      operationId: GithubCallback
      summary: github oauth callback
      description: exchanges code, returns tokens
      tags:
        - Auth
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
  /api/auth/yandex/callback:
    get:
      operationId: YandexCallback
      summary: yandex oauth callback
      description: exchanges code, returns tokens
      tags:
        - Auth
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/auth/refresh:
    post:
      operationId: Refresh
      summary: refresh tokens
      description: exchanges refresh token for new token pair
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
  /api/users/search:
    get:
      operationId: SearchUsers
      summary: search confirmed users
      tags:
        - Users
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchPage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/admin/users:
    get:
      operationId: ListAdminUsers
      summary: list users (admin only)
      tags:
        - Admin
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: confirmed
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersPage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
  /api/admin/users/{id}:
    patch:
      operationId: UpdateAdminUser
      summary: update user (admin only)
      tags:
        - Admin
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserToUpdate'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
  /api/spaces:
    get:
      operationId: ListSpaces
      summary: list spaces
      tags:
        - Spaces
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Space'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
    post:
      operationId: CreateSpace
      summary: create space
      tags:
        - Spaces
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceToCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    IdResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
...
    ErrorMap:
      type: object
      additionalProperties:
        type: string
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        confirmed:
          type: boolean
        confirmed_at:
          type: string
          format: date-time
        confirmed_by:
          type: string
          nullable: true
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time
      additionalProperties: false
    Space:
      type: object
      required:
        - id
        - author_id
        - name
        - created_at
      properties:
        id:
          type: string
        author_id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
...
    SpaceToCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
      additionalProperties: false
...
    UserSearchResult:
      type: object
      required: [id, username, is_admin]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        is_admin:
          type: boolean
      additionalProperties: false
    UserSearchPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserSearchResult'
        next_cursor:
          type: string
      additionalProperties: false
    AdminUsersPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        next_cursor:
          type: string
      additionalProperties: false
    AdminUserToUpdate:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
      additionalProperties: false
...